# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyRepo API",
        "description": "Documentación generada con l5-swagger",
        "contact": {
            "email": "soporte@midominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://myrepo.test",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/admin/businesses": {
            "get": {
                "tags": [
                    "Admin - Businesses"
                ],
                "summary": "Listado de negocios (paginado + filtros).",
                "description": "Filtros:\n- owner_user_id: int\n- active: 1|0\n- search: coincide con name/slug/domain/subdomain/contact_email\n- per_page, page",
                "operationId": "08a21c6e2c87011cd649df8f5d6cb1dc",
                "parameters": [
                    {
                        "name": "owner_user_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "owner_user_id": 10,
                                            "name": "Mi Tienda",
                                            "slug": "mi-tienda",
                                            "domain": "mitienda.com",
                                            "subdomain": "store",
                                            "country_code": "PE",
                                            "currency": "USD",
                                            "timezone": "America/Lima",
                                            "locale": "es_PE",
                                            "contact_name": "Juan",
                                            "contact_email": "soporte@mitienda.com",
                                            "settings": {
                                                "theme": "light"
                                            },
                                            "is_active": true,
                                            "created_at": "2025-08-01T12:00:00Z",
                                            "updated_at": "2025-08-01T12:00:00Z"
                                        }
                                    ],
                                    "meta": {
                                        "current_page": 1,
                                        "last_page": 1,
                                        "per_page": 20,
                                        "total": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Businesses"
                ],
                "summary": "Crear un negocio.",
                "description": "Notas:\n- Si no envías `slug`, se genera desde `name` y se garantiza unicidad.\n- `domain`/`subdomain` se guardan en minúsculas.\n- `currency`/`country_code` se normalizan en mayúsculas.",
                "operationId": "9ff2fb3f60a0d6633e4ae79e38e61d79",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "owner_user_id": {
                                        "type": "integer",
                                        "example": 10,
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Mi Tienda"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "mi-tienda",
                                        "nullable": true
                                    },
                                    "domain": {
                                        "type": "string",
                                        "example": "mitienda.com",
                                        "nullable": true
                                    },
                                    "subdomain": {
                                        "type": "string",
                                        "example": "store",
                                        "nullable": true
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "PE",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "America/Lima",
                                        "nullable": true
                                    },
                                    "locale": {
                                        "type": "string",
                                        "example": "es_PE",
                                        "nullable": true
                                    },
                                    "contact_name": {
                                        "type": "string",
                                        "example": "Juan",
                                        "nullable": true
                                    },
                                    "contact_email": {
                                        "type": "string",
                                        "example": "soporte@mitienda.com",
                                        "nullable": true
                                    },
                                    "settings": {
                                        "type": "object",
                                        "example": {
                                            "theme": "light"
                                        },
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "owner_user_id": 10,
                                    "name": "Mi Tienda",
                                    "slug": "mi-tienda",
                                    "domain": "mitienda.com",
                                    "subdomain": "store",
                                    "country_code": "PE",
                                    "currency": "USD",
                                    "timezone": "America/Lima",
                                    "locale": "es_PE",
                                    "contact_name": "Juan",
                                    "contact_email": "soporte@mitienda.com",
                                    "settings": {
                                        "theme": "light"
                                    },
                                    "is_active": true,
                                    "created_at": "2025-08-01T12:00:00Z",
                                    "updated_at": "2025-08-01T12:00:00Z"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}